fix: break db->layerb->longmemeval->search->db import cycle via internal/aggq leaf package
The merge of main (whose longmemeval now imports internal/search) into the
Layer B branch (whose db imports layerb, which imported longmemeval for the
aggregation classifiers) created an import cycle. Extract the two pure
classifiers and their regexes/stop-words into stdlib-only internal/aggq;
longmemeval keeps thin aliases so its exported API is unchanged.
